seperated into different files for 1)producing data (into csv and txt files) 2)function modules 3)running simulation
