+++ candidate list.txt
@@ -1,0 +1,1 @@
+Bruno Cardi Rocky
