+++ cast votes.csv
@@ -1,0 +1,10 @@
+3.000000000000000000e+00 3.000000000000000000e+00 3.000000000000000000e+00
+5.000000000000000000e+00 4.000000000000000000e+00 1.000000000000000000e+00
+3.000000000000000000e+00 2.000000000000000000e+00 2.000000000000000000e+00
+3.000000000000000000e+00 2.000000000000000000e+00 2.000000000000000000e+00
+2.000000000000000000e+00 3.000000000000000000e+00 3.000000000000000000e+00
+3.000000000000000000e+00 4.000000000000000000e+00 3.000000000000000000e+00
+2.000000000000000000e+00 1.000000000000000000e+00 1.000000000000000000e+00
+2.000000000000000000e+00 4.000000000000000000e+00 1.000000000000000000e+00
+1.000000000000000000e+00 1.000000000000000000e+00 4.000000000000000000e+00
+2.000000000000000000e+00 1.000000000000000000e+00 3.000000000000000000e+00
