+++ kemeny_young.py
@@ -1,29 +1,11 @@
-#!/usr/bin/env python3
-###########shebang line, tells shell what to do with file
-
 #Kemeny-Young method of voting
 #The kemeny-young method is a condorcet method of voting. It satisfies the criterias
 #of unrestricted domain and pareto efficiency but fails independence of irrelevant alternatives
 
-from numpy import *
-import random
 import itertools
-#################################################
-#DEFINE function to process ballots
-def reduce_votes(votes, N_candidates):
 
-    pairwise_ballot=zeros([N_candidates, N_candidates]) #rows represent runners, columns represent opponents
-    for ballot in votes:
-        for a in range(N_candidates):
-            for b in range(N_candidates):
-                if a != b:
-                    if ballot[a]>ballot[b]:
-                        pairwise_ballot[a,b] +=1
-    return(pairwise_ballot, N_candidates) 
-
-#################################################
 #DEFINE function to determine Kemeny-Young ranking
-def Kemeny_Young(pairwise_ballot, N_candidates, candidate_order):
+def kemeny_young(pairwise_ballot, N_candidates, candidate_order):
 
     #score all possible rankings
     all_rankings=list(itertools.permutations(range(N_candidates)))
@@ -61,22 +43,4 @@ def Kemeny_Young(pairwise_ballot, N_candidates, candidate_order):
     print('Final ranking(s) : ', final_ranking_names)
     return(final_ranking_names)
 
-###########################################################################
-###########################################################################
-#imagine that an organization is voting on a candidate to fill a vacancy on its board of directors
-#the candidates' names are Bruno, Cardi, Rocky
-candidate_order = ('Bruno', 'Cardi', 'Rocky') 
-N_candidates = len(candidate_order)
 
-#assign random votes
-N_voters = 10 #number of voters
-votes = zeros([N_voters, N_candidates])
-ballot = zeros(N_candidates)   
-for i in range(N_voters):
-    for a in range (N_candidates):
-        ballot[a]= random.randint(1, 5) 
-    votes[i]=ballot
-    
-#####################################################
-x=reduce_votes(votes, len(candidate_order))
-Kemeny_Young(x[0], x[1], candidate_order)
